requirements.txt
requests
beautifulsoup4
pandas
numpy
torch
transformers
peft
datasets
qdrant-client
sentence-transformers
streamlit
fastapi
uvicorn
openai
python-dotenv
tqdm
evaluate
rouge_score
sacrebleu
firebase-admin
google-cloud-firestore
google-cloud-aiplatform
qdrant-client
sentence-transformers
streamlit
pandas
requests
beautifulsoup4
tqdm
openai
